backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
httpx>=0.28.1
python-dotenv==1.0.0
requests==2.31.0
eventlet==0.34.2
pytest==7.4.3
judge0>=0.0.8
